75-hard-bot/requirements.txt
python-telegram-bot[job-queue]>=20.0,<22.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pytz>=2024.1
Pillow>=10.0.0
anthropic>=0.40.0
